feat: rename to @tableland scope
