+++ package.json
@@ -1,9 +1,8 @@
 {
-  "name": "@textile/tableland",
-  "version": "0.1.2",
+  "name": "@tableland/sdk",
+  "version": "0.0.1",
   "description": "A TypeScript/JavaScript library for creating and querying Tables on the Tableland network.",
-  "author": "Allen Muncy <allen@textile.io>",
-  "repository": "https://github.com/textileio/js-tableland",
+  "repository": "https://github.com/tablelandnetwork/js-tableland",
   "license": "MIT",
   "keywords": [
     "tableland",
@@ -85,8 +84,13 @@
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-typescript": "^7.16.7",
     "@stablelib/base64": "^1.0.1",
-    "@textile/eth-tableland": "0.0.2",
+    "@tableland/eth": "0.0.1",
     "add-js-extension": "^1.0.4",
     "ethers": "^5.5.2"
-  }
+  },
+  "contributors": [
+    "Allen Muncy <allen@textile.io>",
+    "Carson Farmer <carson@textile.io>",
+    "Joe Wagner <joe@textile.io>"
+  ]
 }
